@@ -28,7 +28,7 @@ class UserController extends Controller
         $userFileds['password'] = Hash::make($password);
 
         $user = User::create($userFileds);
-        $user->token = Auth::attempt(['email' => $userFileds['email'], 'password' => $password]);
+        $user->token = auth()->attempt(['email' => $userFileds['email'], 'password' => $password]);
         return response()->json(['user' => $user], 201);
     }
 
@@ -42,7 +42,7 @@ class UserController extends Controller
 
         $userFileds = $request->user;
         $user = User::where('email', '=', $userFileds['email'])->firstOrFail();
-        $user->token = Auth::attempt(['email' => $userFileds['email'], 'password' => $userFileds['password']]);
+        $user->token = auth()->attempt(['email' => $userFileds['email'], 'password' => $userFileds['password']]);
 
         if (!$user->token) {
             return response()->json(['error' => 'Unauthenticated.'], 401);
