@@ -2,17 +2,10 @@
 
 namespace App\Models;
 
-use Illuminate\Auth\Authenticatable;
-use Illuminate\Contracts\Auth\Access\Authorizable as AuthorizableContract;
-use Illuminate\Contracts\Auth\Authenticatable as AuthenticatableContract;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use Laravel\Lumen\Auth\Authorizable;
 
-class Tag extends Model implements AuthenticatableContract, AuthorizableContract
+class Tag extends Model
 {
-    use Authenticatable, Authorizable, HasFactory;
-
     public $timestamps = false;
 
     /**
