@@ -2,18 +2,13 @@
 
 namespace App\Models;
 
-use Illuminate\Auth\Authenticatable;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Database\Eloquent\Casts\Attribute;
-use Illuminate\Contracts\Auth\Access\Authorizable as AuthorizableContract;
-use Illuminate\Contracts\Auth\Authenticatable as AuthenticatableContract;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use Laravel\Lumen\Auth\Authorizable;
+use Tymon\JWTAuth\Contracts\JWTSubject;
+use Illuminate\Notifications\Notifiable;
+use Illuminate\Foundation\Auth\User as Authenticatable;
 
-class User extends Model implements AuthenticatableContract, AuthorizableContract
+class User extends Authenticatable implements JWTSubject
 {
-    use Authenticatable, Authorizable, HasFactory;
+    use Notifiable;
 
     public $timestamps = false;
 
@@ -44,7 +39,7 @@ class User extends Model implements AuthenticatableContract, AuthorizableContrac
 
     protected function getTokenAttribute()
     {
-        return $this->attributes['token'] ?? Auth::guard('api')->getTokenForRequest() ?? null;
+        return $this->attributes['token'] ?? null;
     }
 
     protected function setTokenAttribute($token)
@@ -55,5 +50,25 @@ class User extends Model implements AuthenticatableContract, AuthorizableContrac
     public function articles()
     {
         return $this->hasMany('App\Models\Article');
+    }
+
+        /**
+     * Get the identifier that will be stored in the subject claim of the JWT.
+     *
+     * @return mixed
+     */
+    public function getJWTIdentifier()
+    {
+        return $this->getKey();
+    }
+
+    /**
+     * Return a key value array, containing any custom claims to be added to the JWT.
+     *
+     * @return array
+     */
+    public function getJWTCustomClaims()
+    {
+        return [];
     }
 }
