@@ -1,16 +1,10 @@
 <?php
 
-use Laravel\Lumen\Testing\TestCase as BaseTestCase;
+namespace Tests;
+
+use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
 
 abstract class TestCase extends BaseTestCase
 {
-    /**
-     * Creates the application.
-     *
-     * @return \Laravel\Lumen\Application
-     */
-    public function createApplication()
-    {
-        return require __DIR__.'/../bootstrap/app.php';
-    }
+    use CreatesApplication;
 }
